# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cars API",
    "description": "API for managing cars",
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "https://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/cars": {
      "get": {
        "summary": "Get all cars with pagination",
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cars": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "make": {
                            "type": "string"
                          },
                          "model": {
                            "type": "string"
                          },
                          "year": {
                            "type": "integer"
                          },
                          "price": {
                            "type": "number"
                          },
                          "image": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "make",
                          "model",
                          "year",
                          "price",
                          "image"
                        ]
                      }
                    },
                    "size": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "cars",
                    "size"
                  ]
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cars": {
                      "type": "object",
                      "properties": {
                        "car": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "make": {
                                "type": "string"
                              },
                              "model": {
                                "type": "string"
                              },
                              "year": {
                                "type": "integer"
                              },
                              "price": {
                                "type": "integer"
                              },
                              "image": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "size": {
                      "type": "integer"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new car",
        "tags": [
          "Car"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "make": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  },
                  "mileage": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "integer"
                  },
                  "image": {
                    "type": "string"
                  }
                },
                "required": [
                  "make",
                  "model",
                  "year",
                  "mileage",
                  "price",
                  "image"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Car created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error, car not created"
          },
          "400": {
            "description": "Missing auth header"
          }
        }
      }
    },
    "/cars/{id}": {
      "get": {
        "summary": "Get a specific car by ID",
        "tags": ["Car"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "number"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "make",
                    "model",
                    "year",
                    "mileage",
                    "price",
                    "image",
                    "userId"
                  ]
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Car not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a specific car by ID",
        "tags": [
          "Car"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "make": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "year": {
                    "type": "integer"
                  },
                  "mileage": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "integer"
                  },
                  "image": {
                    "type": "string"
                  }
                },
                "required": [
                  "make",
                  "model",
                  "year",
                  "mileage",
                  "price",
                  "image"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "number"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "make",
                    "model",
                    "year",
                    "price",
                    "image",
                    "userId"
                  ]
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "400": {
            "description": "Missing auth header"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific car by ID",
        "tags": [
          "Car"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "number"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "make",
                    "model",
                    "year",
                    "price",
                    "image",
                    "userId"
                  ]
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "make": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "mileage": {
                      "type": "integer"
                    },
                    "price": {
                      "type": "integer"
                    },
                    "image": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "integer"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "400": {
            "description": "Missing auth header"
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": true,
                    "description": "Min length is 5 char"
                  }
                },
                "required": [
                  "email",
                  "firstName",
                  "lastName",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User has been registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Password should be minimum 5 characters, Email is not valid, Email is already in use",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errorMessage": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An error has occurred, user has not been registered"
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "summary": "Create a new session",
        "tags": [
          "Sessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": true
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": {
                      "type": "string"
                    },
                    "uuid": {
                      "type": "string"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": {
                      "type": "string"
                    },
                    "uuid": {
                      "type": "string"
                    }
                  },
                  "xml": {
                    "name": "root"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "summary": "Delete session when logging out",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "400": {
            "description": "Missing auth header"
          },
          "401": {
            "description": "Session not found"
          },
          "404": {
            "description": "Error"
          }
        }
      }
    }
  }
}